fix: enhance error handling and validation in controllers
- Added error handling for undefined user and pet IDs in createAdoption and deleteUser functions.
- Improved response status codes for successful operations in createAdoption.
- Updated getAllUsers and getUser functions to include error handling for service calls.
- Enhanced Pets.yaml documentation to specify error response structures for 404 and 500 status codes.

+++ RecursosBackend-Adoptme/src/controllers/pets.controller.js
@@ -31,6 +31,7 @@ const updatePet = async(req,res) =>{
 
 const deletePet = async(req,res)=> {
     const petId = req.params.pid;
+    if(petId =='undefined') return res.status(400).send({status:"error",error:"Pet ID is required"})
     const result = await petsService.delete(petId);
     res.send({status:"success",message:"pet deleted"});
 }
